Move encoding helpers to own file
