+++ packages/common/src/utils.ts
@@ -18,15 +18,3 @@ export function removeEmptyFields(obj: {[index: string]:any}): any {
     })
     return obj;
 }
-
-export function base64ToByteArray(blob: string): Uint8Array {
-  return new Uint8Array(
-    atob(blob)
-      .split('')
-      .map((x) => x.charCodeAt(0))
-  );
-}
-
-export function byteArrayToBase64(array: any): string {
-  return btoa(String.fromCharCode.apply(null, array));
-}
